# apps/server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stationary Server API",
    "description": "All API endpoints for the Stationary backend service, supporting export to tools like Apifox / Postman.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9400",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/user/": {
      "get": {
        "summary": "Get current logged-in user info",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string",
                          "nullable": true
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ]
      }
    },
...
    "/api/search/": {
      "get": {
        "summary": "Hybrid search (keyword and vector search)",
        "tags": [
          "Search"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "UNKNOWN",
                "X",
                "XHS",
                "BILIBILI",
                "DOUYIN",
                "TIKTOK",
                "INSTAGRAM"
              ],
              "type": "string"
            }
          },
          {
            "name": "media_type",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "IMAGE",
                "VIDEO",
                "LIVE_PHOTO",
                "AUDIO",
                "PDF"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authenticate requests using API token (Authorization: Bearer <token>)"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Authenticate requests using Better Auth session cookie"
      }
    }
  }
}